sync

+++ utilities/distributed_utilities.py
@@ -1,70 +1,57 @@
 from subprocess import call
-from metadata import *
 import subprocess
 import boto3
 import os
 import sys
 import cPickle as pickle
 import json
-
-sys.path.append(os.environ['REPO_DIR'] + '/utilities')
 from utilities2015 import execute_command
+from metadata import *
 
 def delete_file_or_directory(fp):
     execute_command("rm -rf %s" % fp)
 
-def transfer_data(from_fp, to_fp, from_hostname='localhost', to_hostname='oasis-dm.sdsc.edu'):
+def transfer_data(from_fp, to_fp, from_hostname, to_hostname):
+    assert from_hostname in ['localhost', 'oasis', 's3'], 'from_hostname must be one of localhost, oasis or s3.'
+    assert to_hostname in ['localhost', 'oasis', 's3'], 'to_hostname must be one of localhost, oasis or s3.'
+
     to_parent = os.path.dirname(to_fp)
+    oasis = 'oasis-dm.sdsc.edu'
+
     if from_hostname == 'localhost':
         # upload
-        execute_command("ssh %(to_hostname)s 'rm -rf %(to_fp)s && mkdir -p %(to_parent)s' && scp -r %(from_fp)s %(to_hostname)s:%(to_fp)s" % \
-                    dict(from_fp=from_fp, to_fp=to_fp, to_hostname=to_hostname, to_parent=to_parent))
+        if to_hostname == 's3':
+            execute_command('aws s3 cp --recursive %(from_fp)s s3://%(to_fp)s' % \
+            dict(from_fp=from_fp, to_fp=to_fp))
+        else:
+            assert to_hostname == 'oasis'
+            execute_command("ssh %(to_hostname)s 'rm -rf %(to_fp)s && mkdir -p %(to_parent)s' && scp -r %(from_fp)s %(to_hostname)s:%(to_fp)s" % \
+                    dict(from_fp=from_fp, to_fp=to_fp, to_hostname=oasis, to_parent=to_parent))
     elif to_hostname == 'localhost':
         # download
-        execute_command("rm -rf %(to_fp)s && mkdir -p %(to_parent)s && scp -r %(from_hostname)s:%(from_fp)s %(to_fp)s" % \
-                        dict(from_fp=from_fp, to_fp=to_fp, from_hostname=from_hostname, to_parent=to_parent))
+        if from_hostname == 's3':
+            execute_command('rm -rf %(to_fp)s && mkdir -p %(to_parent)s && aws s3 cp --recursive s3://%(from_fp)s %(to_fp)s' % \
+            dict(from_fp=from_fp, to_fp=to_fp, to_parent=to_parent))
+        else:
+            assert from_hostname == 'oasis'
+            execute_command("rm -rf %(to_fp)s && mkdir -p %(to_parent)s && scp -r %(from_hostname)s:%(from_fp)s %(to_fp)s" % \
+                        dict(from_fp=from_fp, to_fp=to_fp, from_hostname=oasis, to_parent=to_parent))
     else:
         # log onto another machine and perform upload from there.
         execute_command("ssh %(from_hostname)s \"ssh %(to_hostname)s \'rm -rf %(to_fp)s && mkdir -p %(to_parent)s && scp -r %(from_fp)s %(to_hostname)s:%(to_fp)s\'\"" % \
                         dict(from_fp=from_fp, to_fp=to_fp, from_hostname=from_hostname, to_hostname=to_hostname, to_parent=to_parent))
 
-# def upload_to_remote(fp_local, fp_remote, remote_hostname='oasis-dm.sdsc.edu'):
-#     execute_command("scp -r %(fp_local)s %(remote_hostname)s:%(fp_remote)s" % \
-#                     dict(fp_remote=fp_remote, fp_local=fp_local, remote_hostname=remote_hostname))
+default_root = dict(localhost='/home/yuncong', oasis='/home/yuncong/csd395', s3=S3_DATA_BUCKET)
 
-# def download_from_remote(fp_remote, fp_local, remote_hostname='oasis-dm.sdsc.edu'):
-#     execute_command("scp -r %(remote_hostname)s:%(fp_remote)s %(fp_local)s" % \
-#                     dict(fp_remote=fp_remote, fp_local=fp_local, remote_hostname=remote_hostname))
-
-default_root_mapping = dict(localhost='/home/yuncong/CSHL_data_processed', dm='/home/yuncong/csd395/CSHL_data_processed')
-
-def transfer_data_synced(fp_relative, from_hostname='localhost', to_hostname='dm', from_root=None, to_root=None):
+def transfer_data_synced(fp_relative, from_hostname, to_hostname, from_root=None, to_root=None):
     if from_root is None:
-        from_root = default_root_mapping[from_hostname]
+        from_root = default_root[from_hostname]
     if to_root is None:
-        to_root = default_root_mapping[to_hostname]
+        to_root = default_root[to_hostname]
 
     from_fp = os.path.join(from_root, fp_relative)
     to_fp = os.path.join(to_root, fp_relative)
     transfer_data(from_fp=from_fp, to_fp=to_fp, from_hostname=from_hostname, to_hostname=to_hostname)
-
-# def upload_to_remote_synced(fp_relative, stack=None, remote_root='/home/yuncong/csd395/CSHL_data_processed', local_root='/home/yuncong/CSHL_data_processed'):
-#     if stack is not None:
-#         remote_fp = os.path.join(remote_root, stack, fp_relative)
-#         local_fp = os.path.join(local_root, stack, fp_relative)
-#     else:
-#         remote_fp = os.path.join(remote_root, fp_relative)
-#         local_fp = os.path.join(local_root, fp_relative)
-#     create_if_not_exists(os.path.dirname(local_fp))
-#     execute_command("scp -r %(fp_local)s oasis-dm.sdsc.edu:%(fp_remote)s" % \
-#                     dict(fp_remote=remote_fp, fp_local=local_fp))
-
-# def download_from_remote_synced(fp_relative, remote_root='/home/yuncong/csd395/CSHL_data_processed', local_root='/home/yuncong/CSHL_data_processed'):
-#     remote_fp = os.path.join(remote_root, fp_relative)
-#     local_fp = os.path.join(local_root, fp_relative)
-#     create_if_not_exists(os.path.dirname(local_fp))
-#     execute_command("scp -r oasis-dm.sdsc.edu:%(fp_remote)s %(fp_local)s" % \
-#                     dict(fp_remote=remote_fp, fp_local=local_fp))
 
 def first_last_tuples_distribute_over(first_sec, last_sec, n_host):
     secs_per_job = (last_sec - first_sec + 1)/float(n_host)
@@ -74,18 +61,6 @@ def first_last_tuples_distribute_over(first_sec, last_sec, n_host):
         first_last_tuples = [(int(first_sec+i*secs_per_job), int(first_sec+(i+1)*secs_per_job-1) if i != n_host - 1 else last_sec) for i in range(n_host)]
     return first_last_tuples
 
-#def download_dir(client, resource, dist, local='/tmp', bucket='your_bucket'):
-#    paginator = client.get_paginator('list_objects')
-#    for result in paginator.paginate(Bucket=bucket, Delimiter='/', Prefix=dist):
-#        if result.get('CommonPrefixes') is not None:
-#            for subdir in result.get('CommonPrefixes'):
-#                download_dir(client, resource, subdir.get('Prefix'), local)
-#        if result.get('Contents') is not None:
-#            for file in result.get('Contents'):
-#                if not os.path.exists(os.path.dirname(local + os.sep + file.get('Key'))):
-#                     os.makedirs(os.path.dirname(local + os.sep + file.get('Key')))
-#                resource.meta.client.download_file(bucket, file.get('Key'), local + os.sep + file.get('Key'))
-
 def detect_responsive_nodes_aws(exclude_nodes=[], use_nodes=None):
     def get_ec2_avail_instances(region):
         ins = []
@@ -94,16 +69,16 @@ def detect_responsive_nodes_aws(exclude_nodes=[], use_nodes=None):
         response = ec2_conn.describe_instances()
         myid = subprocess.check_output(['wget', '-qO', '-', 'http://instance-data/latest/meta-data/instance-id'])
         for reservation in response["Reservations"]:
-            for instance in reservation["Instances"]: 
-                if instance['State']['Name'] != 'running' or instance['InstanceType'] != 'm4.4xlarge': 
+            for instance in reservation["Instances"]:
+                if instance['State']['Name'] != 'running' or instance['InstanceType'] != 'm4.4xlarge':
                     continue
                 if instance['InstanceId'] != myid:
                     ins.append(instance['PublicDnsName'])
-                else: 
+                else:
                     ins.append('127.0.0.1')
         return ins
     all_nodes = get_ec2_avail_instances('us-west-1')
-    
+
     if use_nodes is not None:
         hostids = use_nodes
     else:
@@ -143,11 +118,47 @@ def detect_responsive_nodes(exclude_nodes=[], use_nodes=None):
 
     return up_hostids
 
-def run_distributed5(command, kwargs_list, stdout=open('/tmp/log', 'ab+'), exclude_nodes=[], use_nodes=None, argument_type='list'):
+def run_distributed(command, kwargs_list, stdout=open('/tmp/log', 'ab+'), exclude_nodes=[], use_nodes=None, argument_type='list', cluster_size=None):
+    if ON_AWS:
+        run_distributed5(command, kwargs_list, cluster_size, stdout, argument_type)
+    else:
+        run_distributed4(command, kwargs_list, stdout, exclude_nodes, use_nodes, argument_type)
+
+
+def run_distributed5(command, kwargs_list, cluster_size, stdout=open('/tmp/log', 'ab+'), argument_type='list'):
+    """
+    Distributed executing a command on AWS.
+    """
+
+    if cluster_size is None:
+        raise Exception('Must specify cluster_size.')
+
+    import time
+
+    n_hosts = subprocess.check_output('qhost').count('\n') - 3
+
+    if n_hosts < cluster_size:
+        autoscaling_description = json.loads(subprocess.check_output('aws autoscaling describe-auto-scaling-groups'.split()))
+        asg = autoscaling_description[u'AutoScalingGroups'][0]['AutoScalingGroupName']
+        subprocess.call("aws autoscaling set-desired-capacity --auto-scaling-group-name %s --desired-capacity %d" % (asg, cluster_size), shell=True)
+        print "Setting autoscaling group %s capaticy to %d\n" % (asg, cluster_size)
+
+        # Wait for SGE to know all nodes. Timeout = 5 mins
+        success = False
+        for _ in range(60):
+            n_hosts = (subprocess.check_output('qhost')).count('\n') - 3
+            if n_hosts == cluster_size:
+                success = True
+                break
+            time.sleep(5)
+
+        if not success:
+            raise Exception('SGE does not receive all host information in 300 seconds. Abort.')
+
+    assert n_hosts == cluster_size
+
     temp_script = '/tmp/runall.sh'
-    #n_hosts = (subprocess.check_output('qhost')).count('\n') - 3
-    n_hosts = 12
-    #3 to remove the header lines
+
     if isinstance(kwargs_list, dict):
         keys, vals = zip(*kwargs_list.items())
         kwargs_list_as_list = [dict(zip(keys, t)) for t in zip(*vals)]
@@ -160,28 +171,50 @@ def run_distributed5(command, kwargs_list, stdout=open('/tmp/log', 'ab+'), exclu
     if argument_type == 'single':
         for arg in kwargs_list_as_list:
             line = command % arg
-    elif argument_type == 'partition':
-        for i, (fi, li) in enumerate(first_last_tuples_distribute_over(0, len(kwargs_list_as_list)-1, n_hosts)):
-        # For cases with partition of first section / last section
-            line = "%(command)s " % \
+    elif argument_type in ['partition', 'list', 'list2']:
+        for i, (fi, li) in enumerate(first_last_tuples_distribute_over(0, len(kwargs_list_as_list)-1, cluster_size)):
+            if argument_type == 'partition':
+                # For cases with partition of first section / last section
+                line = "%(command)s " % \
                     {
                     'command': command % {'first_sec': kwargs_list_as_dict['sections'][fi], 'last_sec': kwargs_list_as_dict['sections'][li]}
                     }
-    elif argument_type == 'list':
-	# Specify kwargs_str
-        for i, (fi, li) in enumerate(first_last_tuples_distribute_over(0, len(kwargs_list_as_list)-1, n_hosts)):
-	    line = "%(command)s " % \
-		    {
-		    'command': command % {'kwargs_str': json.dumps(kwargs_list_as_list[fi:li+1]).replace('"','\\"').replace("'",'\\"')}
-		    }
+            elif argument_type == 'list':
+            # Specify kwargs_str
+                line = "%(command)s " % \
+                {
+                'command': command % {'kwargs_str': json.dumps(kwargs_list_as_list[fi:li+1])}
+                }
+            elif argument_type == 'list2':
+            # Specify {key: list}
+                line = "%(command)s\" &" % \
+                {
+                'command': command % {key: json.dumps(vals[fi:li+1]) for key, vals in kwargs_list_as_dict.iteritems()}
+                }
+            print(line)
+            temp_f = open(temp_script, 'w')
+            temp_f.write(line + '\n')
+            temp_f.close()
+            os.chmod(temp_script, 0o777)
+            call('qsub -V -l mem_free=60G -o %(stdout_log)s -e %(stderr_log)s %(script)s' % \
+                 dict(script=temp_script, stdout_log='/home/ubuntu/stdout.log', stderr_log='/home/ubuntu/stderr.log'),
+                 shell=True, stdout=stdout)
     else:
         raise Exception('argument_type %s not recognized.' % argument_type)
-    print(line)
-    temp_f = open(temp_script, 'w')
-    temp_f.write(line + '\n')
-    temp_f.close()
-    os.chmod(temp_script, 0o777)
-    call('qsub -V -l mem_free=60G ' + temp_script, shell=True, stdout=stdout)
+
+    # Wait for qsub to complete.
+    success = False
+    for _ in range(0, 1200):
+        op = subprocess.check_output('qstat')
+        if "runall.sh" not in op:
+            sys.stderr.write('qsub returned.\n')
+            success = True
+            break
+        time.sleep(5)
+
+    if not success:
+        raise Exception('qsub does not return in 6000 seconds. Abort.')
+
 
 def run_distributed4(command, kwargs_list, stdout=open('/tmp/log', 'ab+'), exclude_nodes=[], use_nodes=None, argument_type='list'):
     """
@@ -191,7 +224,7 @@ def run_distributed4(command, kwargs_list, stdout=open('/tmp/log', 'ab+'), exclu
     hostids = detect_responsive_nodes(exclude_nodes=exclude_nodes, use_nodes=use_nodes)
     print 'Using nodes:', ['gcn-20-%d.sdsc.edu'%i for i in hostids]
     n_hosts = len(hostids)
-    
+
     temp_script = '/tmp/runall.sh'
 
     if isinstance(kwargs_list, dict):
@@ -251,56 +284,3 @@ def run_distributed4(command, kwargs_list, stdout=open('/tmp/log', 'ab+'), exclu
 
     os.chmod(temp_script, 0o777)
     call(temp_script, shell=True, stdout=stdout)
-
-
-
-# def run_distributed3(command, first_sec=None, last_sec=None, interval=1, section_list=None, stdout=None, exclude_nodes=[], take_one_section=True):
-#     """
-#     if take_one_section is True, command must contain %%(secind)d;
-#     otherwise, command must contain %%(f)d %%(l)d
-#     """
-#
-#     hostids = detect_responsive_nodes(exclude_nodes=exclude_nodes)
-#     n_hosts = len(hostids)
-#
-#     temp_script = '/tmp/runall.sh'
-#
-#     # assign a job to each machine
-#     # each line is a job that processes several sections
-#     with open(temp_script, 'w') as temp_f:
-#         if section_list is not None:
-#             for i, (fi, li) in enumerate(first_last_tuples_distribute_over(0, len(section_list)-1, n_hosts)):
-#                 if take_one_section:
-#                     line = "ssh yuncong@%(hostname)s \"%(generic_launcher_path)s \'%(command)s\' --list %(seclist_str)s \" &" % \
-#                             {'hostname': 'gcn-20-%d.sdsc.edu' % hostids[i%n_hosts],
-#                             'generic_launcher_path': os.environ['REPO_DIR'] + '/distributed/generic_launcher.py',
-#                             'command': command,
-#                             # 'seclist_str': '+'.join(map(str, section_list[fi:li+1]))
-#                             'seclist_str': pickle.dumps(section_list[fi:li+1])
-#                             }
-#                     temp_f.write(line + '\n')
-#         else:
-#             for i, (f, l) in enumerate(first_last_tuples_distribute_over(first_sec, last_sec, n_hosts)):
-#                 if take_one_section:
-#                     line = "ssh yuncong@%(hostname)s \"%(generic_launcher_path)s \'%(command)s\' -f %(f)d -l %(l)d -i %(interval)d\" &" % \
-#                             {'hostname': 'gcn-20-%d.sdsc.edu' % hostids[i%n_hosts],
-#                             'generic_launcher_path': os.environ['REPO_DIR'] + '/distributed/generic_launcher.py',
-#                             'command': command,
-#                             'f': f,
-#                             'l': l,
-#                             'interval': interval
-#                             }
-#                 else:
-#                     line = "ssh yuncong@%(hostname)s %(command)s &" % \
-#                             {'hostname': 'gcn-20-%d.sdsc.edu' % hostids[i%n_hosts],
-#                             'command': command % {'f': f, 'l': l}
-#                             }
-#
-#                 temp_f.write(line + '\n')
-#         temp_f.write('wait\n')
-#         temp_f.write('echo =================\n')
-#         temp_f.write('echo all jobs are done!\n')
-#         temp_f.write('echo =================\n')
-#
-#     os.chmod(temp_script, 0o777)
-#     call(temp_script, shell=True, stdout=stdout)
